Move flowide into flowr as a binary
