+++ .github/workflows/build_and_test.yml
@@ -76,7 +76,6 @@ jobs:
         cargo build -p flowc
         cargo build -p flowr
         cargo build -p flowrex
-        cargo build -p flowide
 
     - name: clippy
       run: cargo clippy --tests -- -D warnings
