+++ Cargo.toml
@@ -1,8 +1,7 @@
 [workspace]
-members = ["flowc", "flowstdlib", "flowrlib", "flowr", "flowrex", "flowcore", "flowmacro", "flowsamples",
-    "flowide"]
+members = ["flowc", "flowstdlib", "flowrlib", "flowr", "flowrex", "flowcore", "flowmacro", "flowsamples"]
 default-members = ["flowc", "flowstdlib", "flowrlib", "flowr", "flowrex", "flowcore", "flowmacro",
-    "flowsamples", "flowide"]
+    "flowsamples"]
 resolver = "2"
 
 [workspace.package]
