+++ SUMMARY.md
@@ -45,20 +45,20 @@
   - [Standard Output function](flowr/src/bin/flowrcli/cli/stdio/stdout.md)
   - [Standard Error function](flowr/src/bin/flowrcli/cli/stdio/stderr.md)
 
-# `flowide's` `context functions`
-- [flowide's context functions](flowide/src/gui/flowide_context_functions.md)
-- [Arg functions](flowide/src/gui/args/args.md)
-  - [Arg get function](flowide/src/gui/args/get.md)
-- [File functions](flowide/src/gui/file/file.md)
-  - [File Write function](flowide/src/gui/file/file_write.md)
-  - [File Read function](flowide/src/gui/file/file_read.md)
-- [Image manipulation functions](flowide/src/gui/image/image.md)
-  - [Image buffer](flowide/src/gui/image/image_buffer.md)
-- [Standard IO functions](flowide/src/gui/stdio/stdio.md)
-  - [Readline function](flowide/src/gui/stdio/readline.md)
-  - [Standard Input function](flowide/src/gui/stdio/stdin.md)
-  - [Standard Output function](flowide/src/gui/stdio/stdout.md)
-  - [Standard Error function](flowide/src/gui/stdio/stderr.md)
+# `flowrgui's` `context functions`
+- [flowrgui's context functions](flowr/src/bin/flowrgui/gui/flowrgui_context_functions.md)
+- [Arg functions](flowr/src/bin/flowrgui/gui/args/args.md)
+  - [Arg get function](flowr/src/bin/flowrgui/gui/args/get.md)
+- [File functions](flowr/src/bin/flowrgui/gui/file/file.md)
+  - [File Write function](flowr/src/bin/flowrgui/gui/file/file_write.md)
+  - [File Read function](flowr/src/bin/flowrgui/gui/file/file_read.md)
+- [Image manipulation functions](flowr/src/bin/flowrgui/gui/image/image.md)
+  - [Image buffer](flowr/src/bin/flowrgui/gui/image/image_buffer.md)
+- [Standard IO functions](flowr/src/bin/flowrgui/gui/stdio/stdio.md)
+  - [Readline function](flowr/src/bin/flowrgui/gui/stdio/readline.md)
+  - [Standard Input function](flowr/src/bin/flowrgui/gui/stdio/stdin.md)
+  - [Standard Output function](flowr/src/bin/flowrgui/gui/stdio/stdout.md)
+  - [Standard Error function](flowr/src/bin/flowrgui/gui/stdio/stderr.md)
 
 # Running Flows
 - [Running flows](docs/running/running.md)
