--- flowide/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "flowide"
-description = "A GUI runner for compiled 'flow' programs"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-documentation.workspace = true
-homepage.workspace = true
-repository.workspace = true
-readme = "README.md"
-edition.workspace = true
-
-[dependencies]
-
-flowcore = {path = "../flowcore", version = "0.95.0", features = ["context", "debugger"] }
-flowrlib = {path = "../flowrlib", version = "0.95.0", features = ["context", "debugger", "metrics", "submission"] }
-flowstdlib = {path = "../flowstdlib", version = "0.95.0" }
-clap = "~4"
-log = "0.4.18"
-env_logger = "0.10.0"
-simpath = { version = "~2.4", features = ["urls"]}
-url = { version = "2.2", features = ["serde"] }
-serde_derive = "~1.0.27"
-serde = "~1.0.163"
-serde_json = { version = "1.0", default-features = false }
-error-chain = "0.12.2"
-multimap = "~0.9"
-portpicker = "0.1.1"
-zmq = "0.10.0"
-simpdiscover = "0.6.2"
-iced = { version = "0.9.0", features = ["canvas", "tokio", "debug"] }
-iced_aw = { version = "0.5.2", default-features = false, features = ["tabs", "card", "modal",] }
-once_cell = "1.16.0"
-tokio = { version = "1", features = ["sync"] }
-
-# TODO to remove
-image = "=0.24.6"
-rustyline = "11.0.0"
-
-
-[dev-dependencies]
-tempdir = "~0.3.5"
-serial_test = "2.0.0"
