--- flowide/README.md
@@ -1,3 +1,0 @@
-# FlowIde
-
-Initially just a GUI runtime, later I will add design time stuff
