--- flowide/src/gui/runner.toml
@@ -1,1 +1,0 @@
-name = "flowide"
