+++ flowr/Cargo.toml
@@ -38,8 +38,13 @@ zmq = "0.10.0"
 image = "=0.24.6"
 simpdiscover = "0.6.2"
 
+iced = { version = "0.9.0", features = ["canvas", "tokio", "debug"] }
+iced_aw = { version = "0.5.2", default-features = false, features = ["tabs", "card", "modal",] }
+once_cell = "1.16.0"
+tokio = { version = "1", features = ["sync"] }
+
 # Optional dependencies
-rustyline = {version = "11.0.0", optional = true }
+rustyline = {version = "11.0.0", optional = true } # for debugger
 
 [dev-dependencies]
 tempdir = "~0.3.5"
