+++ flowr/src/bin/flowrgui/errors.rs
