+++ flowr/src/bin/flowrgui/gui/args/get.rs
@@ -1,9 +1,8 @@
 use std::sync::{Arc, Mutex};
 
-use serde_json::{json, Value};
-
 use flowcore::{DONT_RUN_AGAIN, Implementation, RunAgain};
 use flowcore::errors::Result;
+use serde_json::{json, Value};
 
 use crate::gui::client_message::ClientMessage;
 use crate::gui::coordinator_connection::CoordinatorConnection;
@@ -52,11 +51,10 @@ impl Implementation for Get {
 mod test {
     use std::sync::{Arc, Mutex};
 
+    use flowcore::{DONT_RUN_AGAIN, Implementation};
     use portpicker::pick_unused_port;
     use serde_json::json;
     use serial_test::serial;
-
-    use flowcore::{DONT_RUN_AGAIN, Implementation};
 
     use crate::gui::client_message::ClientMessage::Args;
     use crate::gui::coordinator_connection::CoordinatorConnection;
