+++ flowr/src/bin/flowrgui/gui/debug_handler.rs
@@ -1,5 +1,3 @@
-use serde_json::Value;
-
 use flowcore::model::input::Input;
 use flowcore::model::output_connection::OutputConnection;
 use flowcore::model::runtime_function::RuntimeFunction;
@@ -8,6 +6,7 @@ use flowrlib::debug_command::DebugCommand;
 use flowrlib::debugger_handler::DebuggerHandler;
 use flowrlib::job::Job;
 use flowrlib::run_state::{RunState, State};
+use serde_json::Value;
 
 use crate::{BlockBreakpoint, CoordinatorConnection, DataBreakpoint, ExecutionEnded, ExecutionStarted,
             ExitingDebugger, JobCompleted, JobError, Panic, PriorToSendingJob, Resetting,
