+++ flowr/src/bin/flowrgui/gui/debug_message.rs
@@ -1,14 +1,13 @@
 use std::fmt;
 
-use serde_derive::{Deserialize, Serialize};
-use serde_json::Value;
-
 use flowcore::model::input::Input;
 use flowcore::model::output_connection::OutputConnection;
 use flowcore::model::runtime_function::RuntimeFunction;
 use flowrlib::block::Block;
 use flowrlib::job::Job;
 use flowrlib::run_state::{RunState, State};
+use serde_derive::{Deserialize, Serialize};
+use serde_json::Value;
 
 /// A Message sent from the debugger to a debug_client
 #[allow(clippy::large_enum_variant)]
