+++ flowr/src/bin/flowrgui/gui/file/file_write.rs
@@ -1,9 +1,8 @@
 use std::sync::{Arc, Mutex};
 
-use serde_json::Value;
-
 use flowcore::{Implementation, RUN_AGAIN, RunAgain};
 use flowcore::errors::*;
+use serde_json::Value;
 
 use crate::gui::client_message::ClientMessage;
 use crate::gui::coordinator_connection::CoordinatorConnection;
@@ -40,10 +39,9 @@ impl Implementation for FileWrite {
 
 #[cfg(test)]
 mod test {
+    use flowcore::{Implementation, RUN_AGAIN};
     use serde_json::json;
     use serial_test::serial;
-
-    use flowcore::{Implementation, RUN_AGAIN};
 
     use crate::gui::client_message::ClientMessage;
     use crate::gui::coordinator_message::CoordinatorMessage;
