+++ flowr/src/bin/flowrgui/gui/flowrgui_context_functions.md
@@ -1,6 +1,6 @@
-## flowide's Context Functions
+## flowrgui's Context Functions
 Each flow runner can provide a set of functions (referred to as `context functions`) to flows for 
 interacting with the execution environment.
 
-`flowide` is a GUI flow runner, and it provides a set of `context functions` to 
+`flowrgui` is a GUI flow runner, and it provides a set of `context functions` to 
 interact with the file system and standard input/output.
