+++ flowr/src/bin/flowrgui/gui/runner.toml
@@ -1,0 +1,1 @@
+name = "flowrgui"
