+++ flowr/src/bin/flowrgui/gui/stdio/stdout.md
